Привязка конфига и множественные изменения, исправления и доработки.

+++ src/main/java/n2k/nthirst/base/model/ModifierModel.java
@@ -1,6 +1,6 @@
 package n2k.nthirst.base.model;
 public class ModifierModel {
     public float VALUE;
-    public int DURATION;
+    public long DURATION;
     public boolean PERMANENT;
 }
